+++ src/types/ReviewCardProps.ts
@@ -1,0 +1,8 @@
+interface ReviewCardProps {
+  src: string;
+  contents: string;
+  timestamp: string;
+  userId: string;
+}
+
+export default ReviewCardProps;
